Hustlr-backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
black
boto3==1.42.39
botocore==1.42.39
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage
cloudinary
django-cloudinary-storage
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-cors-headers==4.9.0
django-storages==1.14.6
django-redis
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.14
eventlet==0.40.4
greenlet==3.3.1
httpx
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort
jmespath==1.1.0
kombu==5.6.2
packaging==26.0
pillow==12.1.0
pika
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
psycopg-binary==3.3.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
pytest-cov
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
razorpay==1.4.1
redis==7.1.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.3.1
flake8
